src/modules/floating-ip/list-floating-ips/list-floating-ips.spec.ts
import axios from 'axios';
import MockAdapter from 'axios-mock-adapter';
import { createContext } from '../../../utils';
import {listFloatingIps} from './list-floating-ips';
import * as MOCK from './list-floating-ips.mock';

describe('floating-ips', () => {
  const PAGE = 3;
  const PER_PAGE = 26;
  const URL = '/floating_ips';
  const TOKEN = 'bearer-token';
  const mock = new MockAdapter(axios);
  mock.onGet(URL).reply(
    MOCK.response.headers.status,
    MOCK.response.body,
    MOCK.response.headers,
  );
  const context = createContext({
    axios,
    token: TOKEN,
  });
  beforeEach(() => {
    mock.resetHistory();
  });
  describe('list-floating-ips', () => {
    it('should be a fn', () => {
      expect(typeof listFloatingIps).toBe('function');
    });
    it('should return a fn', () => {
      expect(typeof listFloatingIps(context)).toBe('function');
    });
    it('should return IResponse<IListFloatingIpsApiResponse>', async () => {
      const _listFloatingIps = listFloatingIps(context);
      const response = await _listFloatingIps({page: PAGE, per_page: PER_PAGE});
      Object.assign(response, { request: mock.history.get[0]});
      /// validate response schema
      expect(typeof response).toBe('object');
      expect(typeof response.data).toBe('object');
      expect(typeof response.headers).toBe('object');
      expect(typeof response.request).toBe('object');
      expect(typeof response.status).toBe('number');
      /// validate request
      const {request} = response;
      expect(request.url).toBe(context.endpoint + URL);
      expect(request.method).toBe('get');
      expect(request.headers).toMatchObject(MOCK.request.headers);
      expect(request.params).toBeDefined();
      expect(request.params.page).toBe(PAGE);
      expect(request.params.per_page).toBe(PER_PAGE);
      /// validate data
      expect(response.data).toBeDefined();
      expect(response.data.links).toBeDefined();
      expect(response.data.meta).toBeDefined();
      expect(response.data.floating_ips).toBeDefined();
      const {floating_ips} = response.data;
      const [floatingIp] = floating_ips;
      expect(typeof floatingIp.ip).toBe('string');
      /// validate headers
      const {headers, status} = response;
      expect(headers).toMatchObject(MOCK.response.headers);
      expect(status).toBe(MOCK.response.headers.status);
    });
    it('should have default parameters', async () => {
      const defaultPage = 1;
      const defaultper_page = 25;
      const _listFloatingIps = listFloatingIps(context);
      const response = await _listFloatingIps({});
      Object.assign(response, { request: mock.history.get[0]});
      /// validate request
      const {request} = response;
      expect(request.params).toBeDefined();
      expect(request.params.page).toBe(defaultPage);
      expect(request.params.per_page).toBe(defaultper_page);
    });
  });
});
